Rename serializers to serialization
