+++ samples/wearos/src/main/kotlin/com/watchconnection/sample/di/ClientsModule.kt
@@ -1,7 +1,7 @@
 package com.watchconnection.sample.di
 
 import android.content.Context
-import com.boswelja.watchconnection.serializers.StringSerializer
+import com.boswelja.watchconnection.serialization.StringSerializer
 import com.boswelja.watchconnection.wear.discovery.DiscoveryClient
 import com.boswelja.watchconnection.wear.message.MessageClient
 import dagger.Module
