+++ serialization/src/androidMain/kotlin/com/boswelja/watchconnection/serialization/PrimitiveSerializers.kt
@@ -1,4 +1,4 @@
-package com.boswelja.watchconnection.serializers
+package com.boswelja.watchconnection.serialization
 
 import com.boswelja.watchconnection.common.message.MessageSerializer
 import java.nio.ByteBuffer
