--- serializers/src/androidMain/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.boswelja.watchconnection.serializers" />
